pyproject.toml
[tool.poetry]
name = "es-movies"
version = "0.1.0"
description = "Fetch movies from 3rd party API and provide a searchable index with Elasticsearch"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.2"
elasticsearch = {extras = ["async"], version = "^8.12.0"}
dependency-injector = "^4.41.0"
python-dotenv = "^1.0.1"
uvicorn = "^0.27.0.post1"
jinja2 = "^3.1.3"

[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
ipython = "^8.21.0"
ipdb = "^0.13.13"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
